+++ dojo/endpoint/__init__.py
@@ -1,0 +1,1 @@
+import dojo.endpoint.admin  # noqa: F401
