+++ dojo/endpoint/api/__init__.py
@@ -1,0 +1,1 @@
+path = "endpoints"  # noqa: RUF067
