+++ dojo/reports/views.py
@@ -17,9 +17,8 @@ from openpyxl.styles import Font
 from dojo.authorization.authorization import user_has_permission_or_403
 from dojo.authorization.roles_permissions import Permissions
 from dojo.endpoint.queries import get_authorized_endpoints
+from dojo.endpoint.ui.filters import EndpointFilter, EndpointFilterWithoutObjectLookups
 from dojo.filters import (
-    EndpointFilter,
-    EndpointFilterWithoutObjectLookups,
     EndpointReportFilter,
 )
 from dojo.finding.queries import get_authorized_findings
